# What does this PR do?

SD3 pipeline was running softmax kernels with inconsistent tensor shapes leading to low utilization. Observed from the perf lib logs as shown below

`The Geometry of the input and the kernel geometry are inconsistent in softmax_hf8 on account of which the kernel utilization is less than 50 percent  and equals to 25.000000 percent`

### **Optimization :**
This PR addresses the issue of low utilization by padding prompt embeddings to a shape compatible with softmax_hf8 kernels, leading to better utilization and faster execution of diffusion attention process. 

-  Padding prompt embeddings gave a performance increase of 21%

The below table shows the performance change from default execution which uses the regular prompt embedding shape of 333 vs the padded version using 384.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/ddoddbel/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/ddoddbel/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl65
	{font-weight:700;
	text-align:center;}
.xl66
	{font-weight:700;
	text-align:center;}
.xl67
	{text-align:center;}
-->
</head>

<body link="#0563C1" vlink="#954F72">


Batch size | inf steps | Padded Embedding size | No of Images | Batches | Gaudi throughput (samples/sec) | Time for 1 Image Generation (sec)
-- | -- | -- | -- | -- | -- | --
1 | 40 | 384 - padded | 5 | 5 | 0.067 | 14.92537313
1 | 40 | 333 - default | 5| 5| 0.055 | 18.18181818



</body>

</html>


### **Observation from Profile Trace**

The performance gain is achieved because this tensor reshaping reduces the number of softmax kernels used to just one-third of the total compared to the default execution
The information of softmax kernels is summarized below for both default and optimized case

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/ddoddbel/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/ddoddbel/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl65
	{font-weight:700;
	text-align:center;}
.xl66
	{text-align:center;}
-->
</head>

<body link="#0563C1" vlink="#954F72">


Type of   execution | Op Name | Num Nodes | Avg single engine active (us)
-- | -- | -- | --
Default | softmax_fwd_hf8 | 14440 | 145074.07
Default | softmax_fwd_hf8 iter 2 | 14440 | 145071.953
optimized | softmax_fwd_hf8 | 8664 | 145795.544
optimized | softmax_fwd_bf16 | 32 | 1308.537



</body>

</html>



## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?