@deepak-gowda-narayana 2025-03-04
output from make style

ruff check . setup.py --fix
Found 1 error (1 fixed, 0 remaining).
ruff format . setup.py
1 file reformatted, 430 files left unchanged