@deepak-gowda-narayana 2025-03-27
> GaudiStableDiffusion3PipelineTester



> Nice ! Can you check if the SD3 tests all pass please?
> 
> https://github.com/huggingface/optimum-habana/blob/f76f5720c4f9c3d5e9cba54dc0484cd22e9299d4/tests/test_diffusers.py#L1490

All test passes

`================================================================================================================ warnings summary ================================================================================================================
tests/test_diffusers.py::GaudiStableDiffusion3PipelineTester::test_fused_qkv_projections
tests/test_diffusers.py::GaudiStableDiffusion3PipelineTester::test_sd3_inference
tests/test_diffusers.py::GaudiStableDiffusion3PipelineTester::test_stable_diffusion_3_different_negative_prompts
tests/test_diffusers.py::GaudiStableDiffusion3PipelineTester::test_stable_diffusion_3_different_prompts
tests/test_diffusers.py::GaudiStableDiffusion3PipelineTester::test_stable_diffusion_3_prompt_embeds
  /usr/local/lib/python3.10/dist-packages/huggingface_hub/file_download.py:797: FutureWarning: `resume_download` is deprecated and will be removed in version 1.0.0. Downloads always resume when possible. If you want to force a new download, use `force_download=True`.
    warnings.warn(

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=================================================================================================== 5 passed, 5 warnings in 107.04s (0:01:47) ====================================================================================================
`